[TW-940] Change requests per second to batches per second in limits doc (#2156)
* change limit wording in default service limits

* Aligned 429 description with Events API doc

* Edited to make the description of the batches/second limit to match the 429 response description in the Events API doc

* Simplifies explanation of Events API rate limit

Simplifies the explanation of the limit imposed on the Events API so that this information is more accessible to a wider audience. Our goal with this is to explain to users that the thing being throttled here is the number of batches per second, not just the number of requests per second. Some users mistakenly think they can lump multiple batches into a single request to work around this limit, which is false.